pipelines/ingest/requirements.txt
# Ingestion Pipeline Dependencies
# ================================

# HTTP & API
requests>=2.31.0
aiohttp>=3.9.0
httpx>=0.25.0

# Data Processing
pyyaml>=6.0
xmltodict>=0.13.0
python-dateutil>=2.8.2

# Rate Limiting & Retries
tenacity>=8.2.0
ratelimit>=2.2.1
aiolimiter>=1.1.0

# HDFS
hdfs>=2.7.3
pyarrow>=14.0.0

# Logging & Monitoring
structlog>=23.2.0
tqdm>=4.66.0

# Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# CLI
click>=8.1.7
rich>=13.7.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
responses>=0.24.0

